### 2020-11-02 #2799 @will-rowe
Hi @johanneswerner and @jcmcnch. Would either of you be willing to review this submission to JOSS?
