### 2020-11-02 #2799 @jcmcnch
Hi @will-rowe, I would be willing to review Arianna's tool, but in the interests of full disclosure her PhD advisor Harriet is a friend of mine from grad school. Don't know if that presents a conflict of interest from your perspective as an editor. Also FYI I'm also more of a biologist/oceanographer type, so may not offer as useful feedback wrt software design as a true full-blooded bioinformatician (though I am trying to learn this sort of thing myself and am familiar with python). I could be probably most useful simulating the naive user trying to install/use the software and as a skeptical eye towards how well the annotation works. Finally, before I commit fully, is there a timeline that you require? I'm a little busy these days but think I could find time in the coming ~3 weeks to get to this.
