### 2020-11-03 #2799 @will-rowe
Hi @jcmcnch - thanks for getting back to me. I don't think this presents a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).

I think your background makes you a great potential reviewer for JOSS - have a look at the [reviewer criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and see if you think you can assess these points. Usually you don't need to be a bioinformatician/developer to complete a review for us. If you want, I can always help with points you don't feel confident with.

That timeline is fine for me - within a month would be great.

If all this is okay with you and you still want to review for us, please write back and we can start the reviews.
