### 2020-11-04 #2817 @whedon
[REVIEW]: EUKulele: Taxonomic annotation of the unsung eukaryotic microbes
**Submitting author:** @akrinos (<a href="http://orcid.org/0000-0001-9767-8392">Arianna Krinos</a>)
**Repository:** <a href="https://github.com/AlexanderLabWHOI/EUKulele" target ="_blank">https://github.com/AlexanderLabWHOI/EUKulele</a>
**Version:** v1.0.2b
