### 2020-11-16 AlexanderLabWHOI/EUKulele#24 @johanneswerner
Conda installation fails
Hello,

I am trying to install EUKulele as described in the  [documentation](https://eukulele.readthedocs.io/en/latest/quickstart.html#installation) and I am getting the following error:

```
$ conda install -c akrinos -c bioconda -c conda-forge EUKulele
Collecting package metadata (current_repodata.json): done
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: |
Found conflicts! Looking for incompatible packages.
This can take several minutes.  Press CTRL-C to abort.
failed

UnsatisfiableError: The following specifications were found to be incompatible with each other:

Output in format: Requested package -> Available versions
```

I am using a freshly deployed Ubuntu 20.04.1 LTS with conda v. 4.8.3.

Did anyone else have the problem or can reproduce it?
