### 2020-11-16 AlexanderLabWHOI/EUKulele#24 @will-rowe
Hi @johanneswerner . I ran into the same difficulty but think it is because the documentation needs the conda instructions fixing.

In the docs you have:

```
conda create -n EUKulele
conda install -c akrinos -c bioconda -c conda-forge EUKulele
conda activate EUKulele
```

When it should be something like:

```
conda create -n EUKulele
conda activate EUKulele
conda install -c akrinos -c bioconda -c conda-forge EUKulele
```

The current instructions will install the software into the current env, where you may have conflicts etc., and then activate the EUKulele environment. The proposed fix activates a clean environment to use for EUKulele before installing the software. This fixes the issue for me.

As an aside - [mamba](https://github.com/mamba-org/mamba), the drop in conda replacement, massively speeds up the installation for me.
