### 2020-11-16 c5aee67
fix installation commands in quickstart
closes https://github.com/AlexanderLabWHOI/EUKulele/issues/24
