### 2020-11-17 27af709
fixed small typos
