### 2020-11-17 #2817 @johanneswerner
Very interesting software package for the analysis of eukaroytes in metagenomes and metatranscriptomes. I like the focus of this tool and the well-written article and documentation, especially the very comprehensive documentation including all explanations and citations.

I have a few comments that might still be addressed.

- installation
  - [ ] I don't know if this can be improved, but the installation via conda (as described [here](https://eukulele.readthedocs.io/en/latest/install.html)), takes a lot of time.

- documentation
  - [ ] the links **:ref:`documentation`** and **:ref:`Parameters`** are not working in `running-eukulele.rst`
  - [ ] `databaseandconfig.rst`: there are four not three databases

- minimal working example:
  - [ ] the [minimal working example](https://eukulele.readthedocs.io/en/latest/tutorial.html) returns errors with BUSCO

```
EUKulele --config curr_config.yaml

Running EUKulele with entries from the provided configuration file.
No BUSCO file specified/found; using argument-specified organisms and taxonomy for BUSCO analysis.
Setting things up...
Found database folder for reference_DIR in current directory; will not re-download.
Creating a diamond reference from database files...
Aligning to reference database...
['samples_MAGs/sample_2.faa', 'samples_MAGs/sample_1.faa', 'samples_MAGs/sample_0.faa']
Aligning sample sample_2...
Aligning sample sample_1...
Aligning sample sample_0...
Diamond process exited for sample sample_2.
Diamond process exited for sample sample_1.
Diamond process exited for sample sample_0.
Performing taxonomic estimation steps...
Performing taxonomic visualization steps...
Performing taxonomic assignment steps...
Performing BUSCO steps...
Configuring BUSCO...
Running busco with 2 simultaneous jobs...
BUSCO error log:

  File "/home/ubuntu/miniconda3/envs/EUKulele/bin/busco_configurator.py", line 15, in <module>

    for line in open(sys.argv[1]):

FileNotFoundError: [Errno 2] No such file or directory: '/home/ubuntu/.local/bin/../config/config.ini'

sed: can't read test_out_23July/busco/config_sample_1.ini: No such file or directory

sed: can't read test_out_23July/busco/config_sample_1.ini: No such file or directory

sed: can't read test_out_23July/busco/config_sample_1.ini: No such file or directory

ERROR:  Config file test_out_23July/busco/config_sample_2.ini cannot be found

ERROR:  BUSCO analysis failed !

ERROR:  Check the logs, read the user guide, and check the BUSCO issue board on https://gitlab.com/ezlab/busco/issues

BUSCO output log:
python3 busco_configurator.py /home/ubuntu/.local/bin/../config/config.ini test_out_23July/busco/config_sample_1.ini

INFO:   ***** Start a BUSCO v4.1.4 analysis, current time: 11/17/2020 10:19:54 *****

INFO:   Configuring BUSCO with test_out_23July/busco/config_sample_2.ini

BUSCO error log:
ERROR:  Config file test_out_23July/busco/config_sample_1.ini cannot be found

ERROR:  BUSCO analysis failed !

ERROR:  Check the logs, read the user guide, and check the BUSCO issue board on https://gitlab.com/ezlab/busco/issues

BUSCO output log:
INFO:   ***** Start a BUSCO v4.1.4 analysis, current time: 11/17/2020 10:19:54 *****

INFO:   Configuring BUSCO with test_out_23July/busco/config_sample_1.ini

BUSCO error log:

  File "/home/ubuntu/miniconda3/envs/EUKulele/bin/busco_configurator.py", line 15, in <module>

    for line in open(sys.argv[1]):

FileNotFoundError: [Errno 2] No such file or directory: '/home/ubuntu/.local/bin/../config/config.ini'

sed: can't read test_out_23July/busco/config_sample_0.ini: No such file or directory

sed: can't read test_out_23July/busco/config_sample_0.ini: No such file or directory

sed: can't read test_out_23July/busco/config_sample_0.ini: No such file or directory

ERROR:  Config file test_out_23July/busco/config_sample_0.ini cannot be found

ERROR:  BUSCO analysis failed !

ERROR:  Check the logs, read the user guide, and check the BUSCO issue board on https://gitlab.com/ezlab/busco/issues

BUSCO output log:
python3 busco_configurator.py /home/ubuntu/.local/bin/../config/config.ini test_out_23July/busco/config_sample_0.ini

INFO:   ***** Start a BUSCO v4.1.4 analysis, current time: 11/17/2020 10:19:55 *****

INFO:   Configuring BUSCO with test_out_23July/busco/config_sample_0.ini

[] is what is in BUSCO directory
BUSCO initial run did not complete successfully.
Please check the BUSCO run log files in the log/ folder.
```

- tests
  - [ ] `pytest tests/` returns one failed test

```
______________________________________________________ ERROR collecting tests/setupanddownload/test_database.py ______________________________________________________
ImportError while importing test module '/data/EUKulele/tests/setupanddownload/test_database.py'.
/home/ubuntu/miniconda3/envs/EUKulele/lib/python3.6/importlib/__init__.py:126: in import_module
tests/setupanddownload/test_database.py:8: in <module>
    from EUKuleleconfig import *
E   ModuleNotFoundError: No module named 'EUKuleleconfig'
```

- code quality
  - [ ] running pylint (`pylint $(git ls-files '*.py')`) on the repository returns a score of 3.07. I would try to get the pylint score to >=8 (and most warnings are easy to fix).
- minor comments
  - [ ] https://github.com/AlexanderLabWHOI/EUKulele/issues/24
  - [ ] https://github.com/AlexanderLabWHOI/EUKulele/pull/26

- a few comments about the manuscript
  - [ ] How would someone deal with a sample that contains both prokaryotes and eukaryotes? Is it also possible to analyze prokaryotes or is this not desired (if it is possible, how)?
  - [ ] Figure 1: try to avoid arrows that are overlapping the text
  - [ ] The manuscript lists three databases, however in the documentation four databases are listed (EukZoo is missing in the manuscript, see also in `databaseandconfig.rst`)
  - [ ] benchmarking: how long does EUKulele run depending on the size of a dataset/number of sequences? can it be estimated how long a dataset of a certain size will run?
  - [ ] majority of the references do not have dois linked
  - [ ] state of the field: EUKulele is compared to the tools MEGAN and MG-RAST which are as far as I know mostly used for the analyses of prokaryotes (and I believe mostly for metagenomics) - maybe there are better tools to compare EUKulele with
