### 2020-11-17 559a611
Removing old workflow folder and modifying documentation
