+++ workflow/scripts/plotthere.R
@@ -1,5 +1,0 @@
-library(ggplot2)
-library(dplyr)
-library(reshape2)
-
-mets_read = read.csv("METs/SH402_merged-estimated-taxonomy-parallel.out",sep="\t")
