+++ src/EUKulele/EUKulele_main.py
@@ -37,19 +37,19 @@ def main(args_in):
         description='Thanks for using EUKulele! EUKulele is a standalone taxonomic annotation software.\n'
                     'EUKulele is designed primarily for marine microbial eukaryotes. Check the README '
                     'for further information.',
-        usage='eukulele [subroutine] --mets_or_mags [dataset_type] --sample_dir [sample_directory] ' + 
+        usage='eukulele [subroutine] --mets_or_mags [dataset_type] --sample_dir [sample_directory] ' +
               '--reference_dir [reference_database_location] [all other options]')
-    
-    parser.add_argument('subroutine', metavar="subroutine", nargs='?', type=str, default="all", 
-                        choices = ["","all","download","setup","alignment","busco","coregenes"], 
+   
+    parser.add_argument('subroutine', metavar="subroutine", nargs='?', type=str, default="all",
+                        choices = ["","all","download","setup","alignment","busco","coregenes"],
                         help='Choice of subroutine to run.')
-    
+   
     parser.add_argument('-v', '--version', dest = "version", default=False, action='store_true')
-    parser.add_argument('-m', '--mets_or_mags', dest = "mets_or_mags", required = False, default = "") 
-    parser.add_argument('--n_ext', '--nucleotide_extension', dest = "nucleotide_extension", default = ".fasta") 
-    parser.add_argument('--p_ext', '--protein_extension', dest = "protein_extension", default = ".faa") 
+    parser.add_argument('-m', '--mets_or_mags', dest = "mets_or_mags", required = False, default = "")
+    parser.add_argument('--n_ext', '--nucleotide_extension', dest = "nucleotide_extension", default = ".fasta")
+    parser.add_argument('--p_ext', '--protein_extension', dest = "protein_extension", default = ".faa")
     parser.add_argument('-f', '--force_rerun', action='store_true', default=False)
-    parser.add_argument('--scratch', default = '../scratch', 
+    parser.add_argument('--scratch', default = '../scratch',
                         help = "The scratch location to store intermediate files.")
     parser.add_argument('--config_file', default = '')
     parser.add_argument('--perc_mem', dest = "perc_mem", default = 0.75,
@@ -57,46 +57,46 @@ def main(args_in):
 
     ## SALMON OPTIONS ##
     parser.add_argument('--use_salmon_counts', action='store_true', default=False)
-    parser.add_argument('--salmon_dir', 
+    parser.add_argument('--salmon_dir',
                         help = "Salmon directory is required if use_salmon_counts is true.")
     parser.add_argument('--names_to_reads',default=0, help = "A file to be created or used if it exists " +
                         "that relates transcript names to salmon counts from the salmon directory.")
 
     ## WHERE FILES ARE LOCATED ##
-    parser.add_argument('-d', '--database', default="mmetsp", 
+    parser.add_argument('-d', '--database', default="mmetsp",
                         help = "The name of the database to be used to assess the reads.")
-    parser.add_argument('-o', '--out_dir', '--output_dir', dest = "out_dir", default = "output", 
+    parser.add_argument('-o', '--out_dir', '--output_dir', dest = "out_dir", default = "output",
                         help = "Folder where the output will be written.")
     parser.add_argument('-s', '--sample_dir', required = False, dest = "sample_dir", default = "nan",
[…]
