+++ scripts/mag_stats.py
@@ -1,65 +1,76 @@
-#!/usr/bin/env python
-import ujson
-import pandas as pd
-import yaml
-import chardet
-import argparse
-import numpy as np
-import os
+'''
+Function file containing information on how to calculate
+maximum taxonomic level among MAGs.
+'''
 
-level_dict = {'class':['supergroup','division','class'], 'order':['supergroup','division','class', 'order'], 'family':['supergroup','division','class', 'order', 'family'], 'genus': ['supergroup','division','class', 'order', 'family','genus'], 'species':['supergroup','division','class', 'order', 'family','genus', 'species']}
+import os
+import argparse
+import pandas as pd
+import numpy as np
+
+level_dict = {'class':['supergroup','division','class'],
+              'order':['supergroup','division','class','order'],
+              'family':['supergroup','division','class','order','family'],
+              'genus': ['supergroup','division','class','order','family','genus'],
+              'species':['supergroup','division','class','order','family','genus','species']}
 levels = ['supergroup','division','class','order','family','genus','species']
 
 # bug with splitting based on the new taxonomy estimation
 def split_taxonomy(protein_estimate):
+    ''' Split the taxonomy based on the matched protein.'''
     outdf = protein_estimate.copy()
     split_tax_nan = pd.DataFrame(protein_estimate.full_classification.str.split('; '))
     new_col_dict = dict()
-    for l in levels:
-        new_col_dict[str(l)] = [""] * len(outdf.index)
-    for cl in range(len(split_tax_nan.index)):
-        lineage = list(split_tax_nan.loc[:,'full_classification'])[cl]
-        for i,l in enumerate(levels):
+    for curr_level in levels:
+        new_col_dict[str(curr_level)] = [""] * len(outdf.index)
+    for curr_cl in range(len(split_tax_nan.index)):
+        lineage = list(split_tax_nan.loc[:,'full_classification'])[curr_cl]
+        for i,curr_level in enumerate(levels):
             if isinstance(lineage,list):
-                if (len(lineage)>i):
-                    new_col_dict[l][cl] = lineage[i]
+                if len(lineage) > i:
+                    new_col_dict[curr_level][curr_cl] = lineage[i]
                 else:
-                    new_col_dict[l][cl] = np.nan
+                    new_col_dict[curr_level][curr_cl] = np.nan
             else:
-                new_col_dict[l][cl] = np.nan
-    for l in levels:
-        outdf[str(l)] = new_col_dict[l]
-    return(outdf)
+                new_col_dict[curr_level][curr_cl] = np.nan
+    for curr_level in levels:
+        outdf[str(curr_level)] = new_col_dict[curr_level]
+    return outdf
 
 def create_tax_dictionary(split_taxonomy_df):
+    ''' Create a Python dictionary to quickly check taxonomy.'''
     tax_dict = {}
     total_len=len(split_taxonomy_df)
-    for l in levels:
-        tax_dict[l]={}
-        column_sum = split_taxonomy_df.groupby(l)[l].sum()
-        column_count = split_taxonomy_df.groupby(l)[l].count()
+    for curr_level in levels:
+        tax_dict[curr_level]={}
+        # column_sum = split_taxonomy_df.groupby(l)[curr_level].sum()
+        column_count = split_taxonomy_df.groupby(curr_level)[curr_level].count()
         norm_count=column_count/total_len
-        tax_dict[l]=norm_count
-    return(tax_dict)
+        tax_dict[curr_level]=norm_count
+    return tax_dict
 
 def get_max_levels(tax_dict):
+    '''Get the most confident taxonomic assignment.'''
     max_df = pd.DataFrame(index=levels, columns=['max_taxa','percent_id'])
     for key in tax_dict:
[…]
