### 2020-11-18 6db9fe1
Removing defunct tests
