+++ src/EUKulele/EUKulele_main.py
@@ -1,22 +1,9 @@
-#!/usr/bin/env python
+'''
+Software for taxonomic identification of eukaryotes.
+'''
 
-# Software for taxonomic identification of eukaryotes.
-
-import pandas as pd
-import numpy as np
 import os
 import sys
-import yaml
-import chardet
-import argparse
-import multiprocessing
-import subprocess
-import shutil
-import glob
-from joblib import Parallel, delayed
-import pkgutil
-
-import EUKulele
 
 from EUKulele.download_database import downloadDatabase
 from EUKulele.manage_steps import manageEukulele
@@ -24,7 +11,6 @@ from EUKulele.busco_runner import readBuscoFile
 from EUKulele.busco_runner import configRunBusco
 from EUKulele.busco_runner import manageBuscoQuery
 
-import scripts as HelperScripts
 from scripts.names_to_reads import namesToReads
 
 __author__ = "Harriet Alexander, Arianna Krinos"
@@ -39,11 +25,11 @@ def main(args_in):
                     'for further information.',
         usage='eukulele [subroutine] --mets_or_mags [dataset_type] --sample_dir [sample_directory] ' +
               '--reference_dir [reference_database_location] [all other options]')
-   
+  
     parser.add_argument('subroutine', metavar="subroutine", nargs='?', type=str, default="all",
                         choices = ["","all","download","setup","alignment","busco","coregenes"],
                         help='Choice of subroutine to run.')
-   
+  
     parser.add_argument('-v', '--version', dest = "version", default=False, action='store_true')
     parser.add_argument('-m', '--mets_or_mags', dest = "mets_or_mags", required = False, default = "")
     parser.add_argument('--n_ext', '--nucleotide_extension', dest = "nucleotide_extension", default = ".fasta")
@@ -69,7 +55,7 @@ def main(args_in):
                         help = "Folder where the output will be written.")
     parser.add_argument('-s', '--sample_dir', required = False, dest = "sample_dir", default = "nan",
                         help = "Folder where the input data is located (the protein or peptide files to be assessed).")
-   
+  
     ## ONLY SPECIFY THESE ARGUMENTS IF YOU HAVE ALREADY PROVIDED AND FORMATTED YOUR OWN DATABASE ##
     parser.add_argument('--reference_dir', default=".",
                         help = "Folder containing the reference files for the chosen database.")
@@ -77,10 +63,10 @@ def main(args_in):
                         help = "Either a file in the reference directory where the fasta file for the database " +
                                "is located, or a directory containing multiple fasta files that " +
                                "constitute the database.")
- 
+
     parser.add_argument('--tax_table', default = "tax-table.txt")
     parser.add_argument('--protein_map', default = "prot-map.json")
-   
+  
     ## ALIGNMENT OPTIONS ##
     parser.add_argument('--alignment_choice', default = "diamond", choices = ["diamond", "blast"])
 
@@ -111,11 +97,11 @@ def main(args_in):
                        help = "Whether TransDecoder should be run on metatranscriptomic samples. Otherwise, " +
                        "BLASTp is run if protein translated samples are provided, otherwise BLASTx is run " +
                        "on nucleotide samples.")
-   
-   
+  
+  
     parser.add_argument('--test', action='store_true', default=False,
                        help = "Whether we're just running a test and should not execute downloads.")
-              
[…]
