+++ src/EUKulele/busco_runner.py
@@ -1,3 +1,7 @@
+'''
+Facilitates the call to the BUSCO software from within EUKulele.
+'''
+
 import os
 import sys
 import subprocess
@@ -34,7 +38,7 @@ def readBuscoFile(individual_or_summary, busco_file, organisms, organisms_taxono
             organisms = list(busco_file_read.iloc[:,0])
             organisms_taxonomy = list(busco_file_read.iloc[:,1])
             print("Organisms and their taxonomy levels for BUSCO analysis were read from file.")
-           
+          
             if (len(organisms) != len(organisms_taxonomy)):
                 print("Organisms and taxonomic specifications for BUSCO analysis do not contain the same number of entries. " + \
                       "Please revise such that each organism flagged for BUSCO analysis also includes its original " + \
@@ -42,14 +46,14 @@ def readBuscoFile(individual_or_summary, busco_file, organisms, organisms_taxono
                 sys.exit(1)
         else:
             print("No BUSCO file specified/found; using argument-specified organisms and taxonomy for BUSCO analysis.")
-           
+          
     return organisms, organisms_taxonomy
 
 
 def configRunBusco(output_dir, mets_or_mags, pep_ext, nt_ext, sample_dir, samples):
     print("Performing BUSCO steps...", flush=True)
     print("Configuring BUSCO...", flush=True)
-   
+  
     ## Run BUSCO on the full dataset ##
     busco_db = "eukaryota_odb10"
     busco_config_res = configure_busco(busco_db,output_dir)
@@ -73,26 +77,26 @@ def configRunBusco(output_dir, mets_or_mags, pep_ext, nt_ext, sample_dir, sample
         print("BUSCO initial configuration did not complete successfully.\n" +
               "Please check the BUSCO configuration log files in the log/ folder.", flush = True)
         sys.exit(1)
-               
+              
 def configure_busco(busco_db,output_dir):
     busco_config_log = open(os.path.join(output_dir,"log","busco_config.out"), "w+")
     busco_config_err = open(os.path.join(output_dir,"log","busco_config.err"), "w+")
     rc1 = 0
-   
+  
     if not os.path.isdir(os.path.join("busco_downloads","lineages","eukaryota_odb10")):
         p1 = subprocess.Popen(["configure_busco.sh", busco_db], stdout = busco_config_log, stderr = busco_config_err)
         p1.wait()
         rc1 = p1.returncode
     else:
         print("BUSCO lineage database already found; not re-downloaded.")
-       
+      
     busco_config_log.close()
     busco_config_err.close()
     return rc1
-       
+      
 def run_busco(sample_name, output_dir_busco, output_dir, busco_db, mets_or_mags, pep_ext, nt_ext, sample_dir):
     CPUS = multiprocessing.cpu_count()
-   
+  
     if mets_or_mags == "mets":
         if os.path.isfile(os.path.join(output_dir, mets_or_mags, sample_name + "." + pep_ext)):
             fastaname = os.path.join(output_dir, mets_or_mags, sample_name + "." + pep_ext)
@@ -106,36 +110,36 @@ def run_busco(sample_name, output_dir_busco, output_dir, busco_db, mets_or_mags,
     else:
         fastaname = os.path.join(sample_dir, sample_name + "." + pep_ext)
         busco_mode = "proteins"
-       
+      
     busco_run_log = open(os.path.join(output_dir,"log","busco_run.out"), "w+")
     busco_run_err = open(os.path.join(output_dir,"log","busco_run.err"), "w+")
     p1 = subprocess.Popen(["run_busco.sh", str(sample_name), str(output_dir_busco),
                               os.path.join(output_dir_busco, "config_" + sample_name + ".ini"),
[…]
