### 2020-11-18 #2817 @akrinos
@johanneswerner Thank you so much for the very helpful review!

I will respond to what I have responses for thus far!

- Installation: The `conda` installation indeed is quite slow - I am hoping to go through the process of adding it to the `bioconda` channel after publishing the paper, and am hopeful that that will provide a speedup over my user channel.
- Documentation: Thank you for your helpful edits; I have merged those in. Thank you also for pointing out the link issues, which should be fixed now (some of the labels are a bit awkward, which I will fix, but the links I think I converted).
- Minimal working example: I will need to explore your BUSCO error. I have not been working in BUSCO 4.1.4 previously. It may be something where I need to specify that the older version of BUSCO
- Tests: this test is defunct and no longer run by Travis in that form; I have removed it from the tests folder
- Code quality: Working through the formatting issues! Will let you know once I have the numbers up
- I have merged both of the pull requests associated with the suggestions, as well as fixed the documentation inconsistency recommended by another user of the repository

**Other Questions**
- One of our databases, `phylodb`, actually includes prokaryotes, so that is one option, but if prokaryotes were your group of interest, you would probably want to include your own database that is more complete. Beyond that, though, `EUKulele` should work fine on such a sample, although it has specific things built in (e.g. the databases we've chosen) tailored towards eukaryotes
- I will modify the flowchart to fix this and update this thread when that is done
- `EukZoo` was a recent addition; it is not tested on Travis yet, so apologies for the inconsistencies in where it is included!
- Benchmarking is tricky because it's heavily dependent on available memory. However, at a given memory allocation, I can update you with some numbers on estimated runtimes
- Working on linking the DOIs! I saw that in the initial check on the repository
