### 2020-11-18 af97f50
Fixing trailing lines in config
