### 2020-11-18 8245767
Fixing linebreaks for string concatenation
