+++ scripts/query_busco.py
@@ -11,8 +11,7 @@ matches to differentiate between multiple strains.'''
 # python query_busco.py --organism_group Chromera
 # --taxonomic_level genus --output_dir check_dir
 # --fasta_file  --sample_name sample_0
-# --taxonomy_file_prefix   --tax_table " +
-# TAX_TAB + " --busco_out " + busco_table
+# --taxonomy_file_prefix   --tax_table " +\n# TAX_TAB + " --busco_out " + busco_table
 
 import os
 import glob
@@ -87,7 +86,7 @@ def evaluate_organism(organism, taxonomy, tax_table, create_fasta, write_transcr
         #### GET THE CURRENT LEVEL OF TAXONOMY FROM THE TAX TABLE FILE ####
         curr_tax_list = set(list(full_taxonomy[level_hierarchy[curr_level]]))
         if len(curr_tax_list) > 1:
-            print("More than 1 unique match found; using all matches: " +
+            print("More than 1 unique match found; using all matches: " +\
                   str(", ".join(curr_tax_list)), flush=True)
         curr_taxonomy = ";".join(curr_tax_list)
         if (curr_taxonomy == "") | (curr_taxonomy.lower() == "nan"):
@@ -95,8 +94,8 @@ def evaluate_organism(organism, taxonomy, tax_table, create_fasta, write_transcr
             continue
 
         #### CREATE A "MOCK TRANSCRIPTOME" BY PULLING BY TAXONOMIC LEVEL ####
-        taxonomy_file = pd.read_csv(taxonomy_file_prefix + "_all_" +
-                                    str(level_hierarchy[curr_level]) +
+        taxonomy_file = pd.read_csv(taxonomy_file_prefix + "_all_" +\
+                                    str(level_hierarchy[curr_level]) +\
                                     "_counts.csv", sep=",",header=0)
         taxonomy_file = taxonomy_file.loc[[tax in curr_taxonomy for
                                            tax in list(taxonomy_file[level_hierarchy[curr_level].
@@ -147,34 +146,34 @@ def evaluate_organism(organism, taxonomy, tax_table, create_fasta, write_transcr
     reversed_scores.reverse()
 
     if success == 1:
-        file_written = os.path.join(report_dir, level_hierarchy[curr_level] +
+        file_written = os.path.join(report_dir, level_hierarchy[curr_level] +\
                                     "_" + sample_name + ".txt")
         if (write_transcript_file):
             with open(file_written, 'w') as filehandle:
                 for transcript_name in transcripts_to_search_sep:
                     filehandle.write(transcript_name + '\n')
         if (create_fasta):
-            mock_file_name = os.path.join(report_dir, organism + "_" +
-                                          level_hierarchy[curr_level] + "_" +
+            mock_file_name = os.path.join(report_dir, organism + "_" +\
+                                          level_hierarchy[curr_level] + "_" +\
                                           sample_name + "_transcripts.fasta")
-            os.system("grep -w -A 2 -f " + file_written + " " + fasta_file +
+            os.system("grep -w -A 2 -f " + file_written + " " + fasta_file +\
                       " --no-group-separator > " + mock_file_name)
 
-        reported.write("Taxonomy file successfully completed with BUSCO completeness "+
-                       str(busco_completeness) +
-                       "% at location " + str(file_written) +
-                       "\n This was at taxonomic level " +
-                       str(success_level) +
-                       ". \n The file containing the transcript names " +
+        reported.write("Taxonomy file successfully completed with BUSCO completeness "+\
+                       str(busco_completeness) +\
+                       "% at location " + str(file_written) +\
+                       "\n This was at taxonomic level " +\
+                       str(success_level) +\
+                       ". \n The file containing the transcript names " +\
[…]
