+++ src/EUKulele/EUKulele_main.py
@@ -4,6 +4,8 @@ Software for taxonomic identification of eukaryotes.
 
 import os
 import sys
+import argparse
+import multiprocessing
 
 from EUKulele.download_database import downloadDatabase
 from EUKulele.manage_steps import manageEukulele
@@ -19,34 +21,49 @@ __license__ = "MIT"
 __email__ = "akrinos@mit.edu"
 
 def main(args_in):
+    '''
+    Main function for calling subfunctions and running EUKulele.
+    '''
+   
     parser = argparse.ArgumentParser(
-        description='Thanks for using EUKulele! EUKulele is a standalone taxonomic annotation software.\n'
-                    'EUKulele is designed primarily for marine microbial eukaryotes. Check the README '
-                    'for further information.',
-        usage='eukulele [subroutine] --mets_or_mags [dataset_type] --sample_dir [sample_directory] ' +
-              '--reference_dir [reference_database_location] [all other options]')
-  
-    parser.add_argument('subroutine', metavar="subroutine", nargs='?', type=str, default="all",
-                        choices = ["","all","download","setup","alignment","busco","coregenes"],
+        description='Thanks for using EUKulele! EUKulele is a standalone taxonomic '+\
+                    'annotation software.\n EUKulele is designed primarily for marine '+\
+                    'microbial eukaryotes. Check the README for further information.',
+        usage='eukulele [subroutine] --mets_or_mags [dataset_type] --sample_dir ' +\
+              '[sample_directory] --reference_dir [reference_database_location] ' +\
+              '[all other options]')
+
+    parser.add_argument('subroutine', metavar="subroutine", nargs='?', type=str,
+                        default="all",
+                        choices = ["","all","download","setup","alignment",
+                                   "busco","coregenes"],
                         help='Choice of subroutine to run.')
-  
-    parser.add_argument('-v', '--version', dest = "version", default=False, action='store_true')
-    parser.add_argument('-m', '--mets_or_mags', dest = "mets_or_mags", required = False, default = "")
-    parser.add_argument('--n_ext', '--nucleotide_extension', dest = "nucleotide_extension", default = ".fasta")
-    parser.add_argument('--p_ext', '--protein_extension', dest = "protein_extension", default = ".faa")
+
+    parser.add_argument('-v', '--version', dest = "version", default=False,
+                        action='store_true')
+    parser.add_argument('-m', '--mets_or_mags', dest = "mets_or_mags", required = False,
+                        default = "")
+    parser.add_argument('--n_ext', '--nucleotide_extension', dest = "nucleotide_extension",
+                        default = ".fasta")
+    parser.add_argument('--p_ext', '--protein_extension',
+                        dest = "protein_extension",
+                        default = ".faa")
     parser.add_argument('-f', '--force_rerun', action='store_true', default=False)
     parser.add_argument('--scratch', default = '../scratch',
                         help = "The scratch location to store intermediate files.")
     parser.add_argument('--config_file', default = '')
-    parser.add_argument('--perc_mem', dest = "perc_mem", default = 0.75,
-                        help = "The percentage of the total available memory which should be targeted for use by processes.")
+    parser.add_argument('--perc_mem', dest = "perc_mem",
+                        default = 0.75,
+                        help = "The percentage of the total available memory which should "+\
+                               "be targeted for use by processes.")
 
     ## SALMON OPTIONS ##
     parser.add_argument('--use_salmon_counts', action='store_true', default=False)
[…]
