### 2020-11-18 5aee1b2
Fixing some pylint things and again fixing linebreak issues in Travis
