+++ src/EUKulele/busco_runner.py
@@ -10,6 +10,8 @@ import multiprocessing
 from joblib import Parallel, delayed
 import pandas as pd
 
+from scripts.query_busco import queryBusco
+
 MEM_AVAIL_GB = 0
 while MEM_AVAIL_GB == 0:
     try:
@@ -33,8 +35,6 @@ def calc_max_jobs(num_files, size_in_bytes = 2147483648, max_mem_per_proc = 40, 
 
 max_jobs = calc_max_jobs(10)
 
-from scripts.query_busco import queryBusco
-
 def readBuscoFile(individual_or_summary, busco_file, organisms, organisms_taxonomy):
     if individual_or_summary == "individual":
         if (busco_file != "") & (os.path.isfile(busco_file)):
@@ -42,7 +42,7 @@ def readBuscoFile(individual_or_summary, busco_file, organisms, organisms_taxono
             organisms = list(busco_file_read.iloc[:,0])
             organisms_taxonomy = list(busco_file_read.iloc[:,1])
             print("Organisms and their taxonomy levels for BUSCO analysis were read from file.")
-         
+
             if (len(organisms) != len(organisms_taxonomy)):
                 print("Organisms and taxonomic specifications for BUSCO analysis "+\
                       "do not contain the same number of entries. " + \
@@ -53,18 +53,18 @@ def readBuscoFile(individual_or_summary, busco_file, organisms, organisms_taxono
         else:
             print("No BUSCO file specified/found; using argument-specified " +\
                   "organisms and taxonomy for BUSCO analysis.")
-         
+        
     return organisms, organisms_taxonomy
 
 
 def configRunBusco(output_dir, mets_or_mags, pep_ext, nt_ext, sample_dir, samples):
     print("Performing BUSCO steps...", flush=True)
     print("Configuring BUSCO...", flush=True)
- 
+
     ## Run BUSCO on the full dataset ##
     busco_db = "eukaryota_odb10"
     busco_config_res = configure_busco(busco_db,output_dir)
-    n_jobs_busco = min(multiprocessing.cpu_count(), len(samples), 
+    n_jobs_busco = min(multiprocessing.cpu_count(), len(samples),
                        max(1, calc_max_jobs(len(samples))))
     print("Running busco with",n_jobs_busco,"simultaneous jobs...", flush=True)
     busco_res = Parallel(n_jobs=n_jobs_busco, prefer="threads")\
@@ -83,14 +83,14 @@ def configRunBusco(output_dir, mets_or_mags, pep_ext, nt_ext, sample_dir, sample
         print("BUSCO initial configuration did not complete successfully.\n" +
               "Please check the BUSCO configuration log files in the log/ folder.", flush = True)
         sys.exit(1)
-             
+            
 def configure_busco(busco_db,output_dir):
     busco_config_log = open(os.path.join(output_dir,"log",
                                          "busco_config.out"), "w+")
     busco_config_err = open(os.path.join(output_dir,"log",
                                          "busco_config.err"), "w+")
     rc1 = 0
- 
+
     if not os.path.isdir(os.path.join("busco_downloads","lineages",
                                       "eukaryota_odb10")):
         p1 = subprocess.Popen(["configure_busco.sh", busco_db],
@@ -99,11 +99,11 @@ def configure_busco(busco_db,output_dir):
         rc1 = p1.returncode
     else:
         print("BUSCO lineage database already found; not re-downloaded.")
-     
+    
     busco_config_log.close()
     busco_config_err.close()
     return rc1
-     
+    
 def run_busco(sample_name, output_dir_busco, output_dir, busco_db,
[…]
