### 2020-11-19 b4d6596
Adding fixes to syntax
