+++ src/EUKulele/manage_steps.py
@@ -1,12 +1,16 @@
+'''
+EUKulele step manager.
+'''
+
 import os
 import sys
 import subprocess
 import multiprocessing
-from joblib import Parallel, delayed
+import math
 import shutil
 import pathlib
+from joblib import Parallel, delayed
 import pandas as pd
-import math
 
 import EUKulele
 from EUKulele.tax_placement import place_taxonomy
@@ -14,28 +18,28 @@ from EUKulele.visualize_results import visualize_all_results
 
 from scripts.mag_stats import magStats
 
-MEM_AVAIL_GB = 0
-while MEM_AVAIL_GB == 0:
+mem_avail_gb = 0
+while mem_avail_gb == 0:
     try:
         os.system("free -m > free.csv")
-        MEM_AVAIL_GB = pd.read_csv("free.csv", sep = "\s+").free[0] / 10**3
+        mem_avail_gb = pd.read_csv("free.csv", sep = "\s+").free[0] / 10**3
     except:
         pass
-   
+
 # 25 GB memory per GB file size
 # add a parameter that changes between the requirement per file size (reducing to 10 GB for now)
 # also add a parameter to EUKulele that decides whether you use 100% of available memory and scales
-# MEM_AVAIL_GB by that amount (default to 75%)
+# mem_avail_gb by that amount (default to 75%)
 def calc_max_jobs(num_files, size_in_bytes = 2147483648, max_mem_per_proc = 10, perc_mem = 0.75):
     size_in_gb = size_in_bytes / (1024*1024*1024)
     if size_in_gb == 0:
         size_in_gb = 0.01
-    MAX_JOBS = math.floor(MEM_AVAIL_GB * perc_mem / (max_mem_per_proc * size_in_gb * num_files)) #48)
-    if MAX_JOBS == 0:
-        MAX_JOBS = 1
-    return MAX_JOBS
-       
-MAX_JOBS = max(1, calc_max_jobs(1))
+    max_jobs = math.floor(mem_avail_gb * perc_mem / (max_mem_per_proc * size_in_gb * num_files)) #48)
+    if max_jobs == 0:
+        max_jobs = 1
+    return max_jobs
+
+max_jobs = max(1, calc_max_jobs(1))
 
 # For DIAMOND: The program can be expected to use roughly six times this number of memory (in GB).
 # So for the default value of -b2.0, the memory usage will be about 12 GB.
@@ -43,31 +47,34 @@ MAX_JOBS = max(1, calc_max_jobs(1))
 
 def manageEukulele(piece, mets_or_mags = "", samples = [], database_dir = "",
                    output_dir = "", ref_fasta = "", alignment_choice = "diamond",
-                   rerun_rules = False, cutoff_file = "", sample_dir = "", nt_ext = "", pep_ext = "",
-                   consensus_cutoff = 0.75, tax_tab = "", prot_tab = "", use_salmon_counts = False,
+                   rerun_rules = False, cutoff_file = "", sample_dir = "",
+                   nt_ext = "", pep_ext = "",consensus_cutoff = 0.75,
+                   tax_tab = "", prot_tab = "", use_salmon_counts = False,
                    names_to_reads = "", alignment_res = "", filter_metric = "evalue",
                    run_transdecoder = False, transdecoder_orf_size = 100, perc_mem = 0.75):
-   
+
     """
     This function diverts management tasks to the below helper functions.
     """
-   
+
     if piece == "setup_eukulele":
         setupEukulele(output_dir)
     elif piece == "setup_databases":
-        createAlignmentDatabase(ref_fasta, rerun_rules, output_dir, alignment_choice, database_dir)
+        createAlignmentDatabase(ref_fasta, rerun_rules, output_dir,
[…]
