+++ src/EUKulele/visualize_results.py
@@ -11,23 +11,31 @@ import argparse
 #.loc[[name == curr for name in final_frame.loc[name_level]],["Sum"]]
 
 def countClassifs(level, level_hierarchy, name_level, df):
+    ''' Count the classifications assigned to each level and type. '''
+
     set_list = set()
-   
+
     classifications = list(df.loc[df["classification_level"] == level]["classification"])
     counts = list(df.loc[df["classification_level"] == level]["counts"])
     transcript_names = list(df.loc[df["classification_level"] == level]["transcript_name"])
     match_loc = int(np.where([curr == level for curr in level_hierarchy])[0])
     for curr in range(match_loc + 1,len(level_hierarchy)):
-        classification_curr = list(df.loc[df["classification_level"] == level_hierarchy[curr]]["full_classification"])
-        transcripts_curr = list(df.loc[df["classification_level"] == level_hierarchy[curr]]["transcript_name"])
-        correct_index = list(np.where([len(str(cr).split(";")) >= abs(-1-(curr-match_loc)) for cr in classification_curr])[0])
-       
+        classification_curr = list(df.loc[df["classification_level"] == \
+                                          level_hierarchy[curr]]["full_classification"])
+        transcripts_curr = list(df.loc[df["classification_level"] == \
+                                       level_hierarchy[curr]]["transcript_name"])
+        correct_index = list(np.where([len(str(cr).split(";")) >= \
+                                       abs(-1-(curr-match_loc)) \
+                                       for cr in classification_curr])[0])
+
         classification_curr = [classification_curr[cr2] for cr2 in correct_index]
-        classifs_curr = [str(cr).split(";")[match_loc].strip() for cr in classification_curr]
+        classifs_curr = [str(cr).split(";")[match_loc].strip() \
+                         for cr in classification_curr]
         transcripts_curr = [transcripts_curr[cr2] for cr2 in correct_index]
-        counts_curr = list(df.loc[df["classification_level"] == level_hierarchy[curr]]["counts"])
+        counts_curr = list(df.loc[df["classification_level"] == \
+                                  level_hierarchy[curr]]["counts"])
         counts_curr = [counts_curr[cr2] for cr2 in correct_index]
-       
+
         # add to running list
         classifications.extend(classifs_curr)
         transcript_names.extend(transcripts_curr)
@@ -36,83 +44,108 @@ def countClassifs(level, level_hierarchy, name_level, df):
     counts.extend(list(lostducklings.counts))
     transcript_names.extend(list(lostducklings.transcript_name))
     classifications.extend(["NoClassification"] * len(lostducklings.index))
-       
+      
     classifications = [str(cr).strip().strip(""''"]['") for cr in classifications]
     full_list = classifications
     set_list.update(set(classifications))
-   
-    transcripts_classes = pd.DataFrame({"classifications": classifications, "transcript_names": transcript_names})
-    transcripts_classes = transcripts_classes.groupby("classifications").agg({"transcript_names": lambda x: ';'.join(x)})
-   
+
+    transcripts_classes = pd.DataFrame({"classifications": classifications,\
+                                        "transcript_names": transcript_names})
+    transcripts_classes = transcripts_classes.groupby("classifications").agg(\
+        {"transcript_names": lambda x: ';'.join(x)})
+
     # Apparently now when you groupby, the grouping becomes the index.
     if transcripts_classes.index.name != "classifications":
         transcripts_classes = transcripts_classes.set_index('classifications')
     transcripts_classes = transcripts_classes.loc[sorted(list(set_list))]
-   
+
[…]
