### 2020-11-19 ed1a04c
Fixing bug in alignment_choice variable
