### 2020-11-19 7153bf1
Fixing bug in alignment_choice variable
