+++ tests/code/test_mags_2.py
@@ -12,6 +12,11 @@ import os
 test_reference = "mmetsp"
 
 def test_individual():
+    '''
+    Test running the BUSCO functionality with individually specified
+    arguments, rather than a full class/other functional group.
+    '''
+
     base_dir = os.path.join(os.path.dirname(__file__), '..', 'aux_data')
     sample_dir = os.path.join(base_dir, test_reference, "samples_MAGs")
     output_dir = os.path.join(base_dir, "test_out_E")
@@ -30,12 +35,16 @@ def test_individual():
     assert os.path.isfile(busco_out)
 
 def test_error_input():
+    '''
+    Checking that we can catch an expected error in the input formatting.
+    '''
+
     base_dir = os.path.join(os.path.dirname(__file__), '..', 'aux_data')
     sample_dir = os.path.join(base_dir, test_reference, "samples_MAGs")
     output_dir = os.path.join(base_dir, "test_out_F")
     os.system("rm -rf " + output_dir)
     reference_dir = os.path.join(base_dir, test_reference, "sample_ref_MAGs")
-  
+
     string_arguments = " ".join(["--database", "mmetsp", "--sample_dir", sample_dir,
                                  "--mets_or_mags", "mmm", "--out_dir", output_dir, "-i",
                                  '--organisms', 'Chromera', '--taxonomy_organisms', 'genus',
@@ -45,16 +54,20 @@ def test_error_input():
         eukulele(string_arguments=string_arguments)
     except:
         error = 1
-  
+
     assert error == 1
 
 def test_error_required_input():
+    '''
+    Tests that we receive an error when a required input is not included.
+    '''
+
     base_dir = os.path.join(os.path.dirname(__file__), '..', 'aux_data')
     sample_dir = os.path.join(base_dir, test_reference, "samples_MAGs")
     output_dir = os.path.join(base_dir, "test_out_G")
     os.system("rm -rf " + output_dir)
     reference_dir = os.path.join(base_dir, test_reference, "sample_ref")
-  
+
     string_arguments = " ".join(["--database", "mmetsp", "--sample_dir", sample_dir,
                                  "--out_dir", output_dir, "-i",
                                  '--organisms', 'Chromera', '--taxonomy_organisms', 'genus',
@@ -64,16 +77,20 @@ def test_error_required_input():
         eukulele(string_arguments=string_arguments)
     except:
         error = 1
-  
+
     assert error == 1
 
 def test_error_busco_no_orgs():
+    '''
+    Test that we get an error when we do not specify an organism.
+    '''
+
     base_dir = os.path.join(os.path.dirname(__file__), '..', 'aux_data')
     sample_dir = os.path.join(base_dir, test_reference, "samples_MAGs")
     output_dir = os.path.join(base_dir, "test_out_H")
     os.system("rm -rf " + output_dir)
     reference_dir = os.path.join(base_dir, test_reference, "sample_ref")
-  
+
     string_arguments = " ".join(["--database", "mmetsp", "--sample_dir", sample_dir,
                                  "--mets_or_mags", "mags", "--out_dir", output_dir, "-i",
                                  "--reference_dir", reference_dir])
@@ -82,16 +99,21 @@ def test_error_busco_no_orgs():
         eukulele(string_arguments=string_arguments)
     except:
         error = 1
-  
+
     assert error == 1
 
 def test_error_n_extension():
+    '''
+    Tests that we get an error when we use an improper nucleotide
+    extension for the files.
+    '''
+
     base_dir = os.path.join(os.path.dirname(__file__), '..', 'aux_data')
     sample_dir = os.path.join(base_dir, test_reference, "samples_MAGs")
     output_dir = os.path.join(base_dir, "test_out_I")
[…]
