+++ src/EUKulele/manage_steps.py
@@ -72,33 +72,38 @@ def manageEukulele(piece, mets_or_mags = "", samples = [], database_dir = "",
                              pep_ext.strip('.'),
                      run_transdecoder = run_transdecoder, perc_mem = perc_mem)
     elif piece == "align_to_db":
-        return manageAlignment(alignment_choice, samples, filter_metric, output_dir, ref_fasta,
-                               mets_or_mags, database_dir, sample_dir, rerun_rules,
-                               nt_ext, pep_ext, core = "full",perc_mem = perc_mem)
+        return manageAlignment(alignment_choice, samples, filter_metric,
+                               output_dir, ref_fasta,mets_or_mags, database_dir,
+                               sample_dir, rerun_rules, nt_ext, pep_ext,
+                               core = "full",perc_mem = perc_mem)
     elif piece == "estimate_taxonomy":
-        manageTaxEstimation(output_dir, mets_or_mags, tax_tab, cutoff_file, consensus_cutoff,
-                            prot_tab, use_salmon_counts, names_to_reads, alignment_res,
-                            rerun_rules, samples, sample_dir, pep_ext, nt_ext, perc_mem)
+        manageTaxEstimation(output_dir, mets_or_mags, tax_tab, cutoff_file,
+                            consensus_cutoff, prot_tab, use_salmon_counts,
+                            names_to_reads, alignment_res, rerun_rules, samples,
+                            sample_dir, pep_ext, nt_ext, perc_mem)
     elif piece == "visualize_taxonomy":
         manageTaxVisualization(output_dir, mets_or_mags, sample_dir, pep_ext, nt_ext,
                                use_salmon_counts, rerun_rules)
     elif piece == "assign_taxonomy":
         manageTaxAssignment(samples, mets_or_mags, output_dir, sample_dir, pep_ext, core = False)
     elif piece == "core_align_to_db":
-        alignment_res = manageAlignment(alignment_choice, samples, filter_metric, output_dir, ref_fasta,
-                        mets_or_mags, database_dir, sample_dir, rerun_rules, nt_ext, pep_ext, core = "core",
-                                       perc_mem = perc_mem)
+        alignment_res = manageAlignment(alignment_choice, samples, filter_metric,
+                                        output_dir, ref_fasta, mets_or_mags, database_dir,
+                                        sample_dir, rerun_rules, nt_ext, pep_ext, core = "core",
+                                        perc_mem = perc_mem)
         alignment_res = [curr for curr in alignment_res if curr != ""]
         return alignment_res
     elif piece == "core_estimate_taxonomy":
-        manageCoreTaxEstimation(output_dir, mets_or_mags, tax_tab, cutoff_file, consensus_cutoff,
-                            prot_tab, use_salmon_counts, names_to_reads, alignment_res,
-                            rerun_rules, samples, sample_dir, pep_ext, nt_ext, perc_mem)
+        manageCoreTaxEstimation(output_dir, mets_or_mags, tax_tab, cutoff_file,
+                                consensus_cutoff, prot_tab, use_salmon_counts,
+                                names_to_reads, alignment_res, rerun_rules, samples,
+                                sample_dir, pep_ext, nt_ext, perc_mem)
     elif piece == "core_visualize_taxonomy":
         manageCoreTaxVisualization(output_dir, mets_or_mags, sample_dir, pep_ext, nt_ext,
                                use_salmon_counts, rerun_rules, core = True)
     elif piece == "core_assign_taxonomy":
-        manageTaxAssignment(samples, mets_or_mags, output_dir, sample_dir, pep_ext, core = True)
[…]

[…]
