### 2020-11-20 eed9618
Fixing removal of the last character in pytest
