### 2020-11-20 034d05f
Trying modifying the reference directory to fix race condition
