### 2020-11-22 e9ae414
Adding DOIs and updating paper to list all 4 database options
