+++ paper/paper.bib
@@ -75,6 +75,7 @@
   title={Metagenomic taxonomic classification using extreme learning machines},
   author={Rasheed, Zeehasham and Rangwala, Huzefa},
   journal={Journal of bioinformatics and computational biology},
+  doi={10.1142/S0219720012500151},
   volume={10},
   number={05},
   pages={1250015},
@@ -86,6 +87,7 @@
   title={Fast and sensitive taxonomic classification for metagenomics with Kaiju},
   author={Menzel, Peter and Ng, Kim Lee and Krogh, Anders},
   journal={Nature communications},
+  doi={10.1038/ncomms11257},
   volume={7},
   number={1},
   pages={1--9},
@@ -97,6 +99,7 @@
   title={A metagenomic assessment of microbial eukaryotic diversity in the global ocean},
   author={Obiol, Aleix and Giner, Caterina R and S{\'a}nchez, Pablo and Duarte, Carlos M and Acinas, Silvia G and Massana, Ramon},
   journal={Molecular Ecology Resources},
+  doi={10.1111/1755-0998.13147},
   volume={20},
   number={3},
   year={2020},
@@ -107,6 +110,7 @@
   title={Metatranscriptomic census of active protists in soils},
   author={Geisen, Stefan and Tveit, Alexander T and Clark, Ian M and Richter, Andreas and Svenning, Mette M and Bonkowski, Michael and Urich, Tim},
   journal={The ISME journal},
+  doi={10.1038/ismej.2015.30},
   volume={9},
   number={10},
   pages={2178--2190},
@@ -119,6 +123,7 @@
   author={Caron, David A and Alexander, Harriet and Allen, Andrew E and Archibald, John M and Armbrust, E Virginia and Bachy, Charles and Bell, Callum J and Bharti, Arvind and Dyhrman, Sonya T and Guida, Stephanie M and others},
   journal={Nature Reviews Microbiology},
   volume={15},
+  doi={10.1038/nrmicro.2016.160},
   number={1},
   pages={6--20},
   year={2017},
@@ -128,6 +133,7 @@
 @article{yu2020sager,
   title={SAGER: a database of Symbiodiniaceae and Algal Genomic Resource},
   author={Yu, Liying and Li, Tangcheng and Li, Ling and Lin, Xin and Li, Hongfei and Liu, Chichi and Guo, Chentao and Lin, Senjie},
+  doi={10.1093/database/baaa051},
   journal={Database},
   volume={2020},
   year={2020},
@@ -137,21 +143,13 @@
 @article{price2017robust,
   title={Robust Dinoflagellata phylogeny inferred from public transcriptome databases},
   author={Price, Dana C and Bhattacharya, Debashish},
-  journal={Journal of phycology},
+  journal={Journal of Phycology},
+  doi={10.1111/jpy.12529},
   volume={53},
   number={3},
   pages={725--729},
   year={2017},
   publisher={Wiley Online Library}
-}
-
-@article{yu2020sager,
-  title={SAGER: a database of Symbiodiniaceae and Algal Genomic Resource},
-  author={Yu, Liying and Li, Tangcheng and Li, Ling and Lin, Xin and Li, Hongfei and Liu, Chichi and Guo, Chentao and Lin, Senjie},
-  journal={Database},
-  volume={2020},
-  year={2020},
-  publisher={Oxford Academic}
 }
 
 @article{kranzler2019silicon,
@@ -160,6 +158,7 @@
   journal={Nature microbiology},
   volume={4},
   number={11},
+  doi={10.1038/s41564-019-0502-x},
   pages={1790--1797},
   year={2019},
   publisher={Nature Publishing Group}
@@ -169,6 +168,7 @@
[…]
