### 2020-11-23 0f9b86f
Fixing problem in bib file and increasing pylint score
