+++ paper/paper.bib
@@ -457,7 +457,7 @@
 @incollection{richard2020eukaryotic,
   title={Eukaryotic Pangenomes},
   author={Richard, Guy-Franck},
-  doi={10.1007/978-3-030-38281-0_12}
+  doi={10.1007/978-3-030-38281-0_12},
   booktitle={The Pangenome},
   pages={253--291},
   year={2020},
