### 2020-11-23 #2817 @akrinos
As an update to the above:

- DOIs are now included for all references for which DOIs are available
- The score returned by `pylint $(git ls-files '*.py')` is now 8.26/10
- The flowchart arrows have been moved such that they are no longer covering text on the documentation landing page, as well as in the paper
- All four default databases are now referenced in the paper

We are working on benchmarking and addressing `BUSCO`-related issues. Thank you for your patience!
