### 2020-11-23 0e1a9b4
Updating conda recipe and EUKulele configuration
