### 2020-11-23 #2817 @akrinos
@johanneswerner I was able to reproduce the error you have been getting from `BUSCO` when trying to run the `sample_EUKulele` example folder.

It is indeed related to `BUSCO` version 4.1.4, in which version the `BUSCO` configuration file is stored in a different location than it was previously. I have implemented a patch that has been deployed to the `conda` build of `EUKulele` that searches for the configuration file in a different way. This solves the issue of the initial `BUSCO` run, but the storage location of the final `BUSCO` sequences will still be different due to the version change. For now, `EUKulele` can be run with `BUSCO` version 4.0.6 and `Biopython` 1.77. I will sort out the final versioning issues on the `pip` install version such that both versions should work; for now at least the `BUSCO` run itself functions.

Thanks again for your patience!
