### 2020-11-24 6915c56
Fixing issues with BUSCO for newer versions
