### 2020-11-25 #2817 @akrinos
@johanneswerner we have provided a graphic below using the `DIAMOND` alignment tool for various sizes of sequence files. The time is in minutes in terms of how long the full `EUKulele` run takes to execute. Note that this is for metatranscriptomic (MET) sequences with or without using the `TransDecoder` tool for translation (as well as colored by two different database selections, the MMETSP and PhyloDB).

We have also uploaded `pip` and `conda`-installable revisions of `EUKulele` which address the issue you encountered with the latest `BUSCO` version. In recently-uploaded version 1.0.1, these issues should be resolved, and you should be able to fully execute the small test example which you reported a prior test of above.

Thank you!

![image](https://user-images.githubusercontent.com/22385511/100262472-2e74e500-2f1a-11eb-9030-8d17df28355d.png)
