### 2020-11-26 #2817 @johanneswerner
Dear @akrinos

thank you for your updates. I think I checked the above checkboxes that are taken care of (if I forgot something, please let me know).

Unfortunately, I still encountered errors with the minimum example ([run.log](https://github.com/openjournals/joss-reviews/files/5603251/run.log)) and some of the tests also throw errors on my virtual instance ([tests.log](https://github.com/openjournals/joss-reviews/files/5603254/tests.log)). Could you please have a look at them?

Thank you very much for your effort, especially the benchmarking is very interesting.
