### 2020-11-26 #2817 @akrinos
Hmm, it looks like you're still getting the same error, which is most likely the cause for the failed tests as well (although I haven't looked carefully at each failure). Did you reinstall via `conda` @johanneswerner ? It looks like from the error that it is defaulting to using the `BUSCO` install that you have locally, rather than a `BUSCO` install via `conda`. Could you please try running `EUKulele --version`? The problem is also in the included scripts `run_busco.sh` and `concatenate_busco.sh`, so printing the result of `cat $(which run_busco.sh)` and `cat $(which concatenate_busco.sh)` to a file would also help me verify that the fix that I have added is present in the files that your install is pulling. One problem I had was needing to remove prior installs.

If this continues to be an issue, I suppose we should move to another thread per the guidelines. Thanks for your persistence!
