### 2020-11-26 #2817 @johanneswerner
My apologies @akrinos, I pulled the git repository for the tests, but I forgot to reinstall via `conda`. Thank you for looking into it. :-)

Test dataset runs accordingly after reinstallation with `conda`, and the tests also pass. I marked the respective checkboxes above.
