### 2020-12-03 AlexanderLabWHOI/EUKulele#24 @jcmcnch
@akrinos, this still seems to be an issue. I can't install - getting same error as Johannes with both conda and mamba.

conda error:

```Collecting package metadata (current_repodata.json): done
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: -
Found conflicts! Looking for incompatible packages.
This can take several minutes.  Press CTRL-C to abort.
failed

UnsatisfiableError: The following specifications were found to be incompatible with each other:

Output in format: Requested package -> Available versions
Note that strict channel priority may have removed packages required for satisfiability.
```

mamba error:
```
Looking for: ['eukulele']

akrinos/linux-64         [====================] (00m:00s) Done
akrinos/noarch           [====================] (00m:00s) Done
ursky/linux-64           [====================] (00m:00s) Done
pkgs/main/noarch         [====================] (00m:01s) Done
ursky/noarch             [====================] (00m:00s) Done
bioconda/noarch          [====================] (00m:03s) Done
pkgs/r/noarch            [====================] (00m:00s) Done
bioconda/linux-64        [====================] (00m:04s) Done
conda-forge/noarch       [====================] (00m:04s) Done
pkgs/main/linux-64       [====================] (00m:04s) Done
pkgs/r/linux-64          [====================] (00m:04s) Done
conda-forge/linux-64     [====================] (00m:15s) Done
Encountered problems while solving.
Problem: package eukulele-0.0.1-pyh39e3cac_0 requires argparse, but none of the providers can be installed
```
