### 2020-12-03 AlexanderLabWHOI/EUKulele#24 @johanneswerner
@jcmcnch

Good morning,

I have an idea what could have caused this error. Could you share your bash commands for setting up the environment? I would like to test if I could reproduce it.
