### 2020-12-03 AlexanderLabWHOI/EUKulele#24 @akrinos
Hi @jcmcnch, thanks for bringing this up. I don't see too many issues referencing the error online, though it's coming from Linux. I also note here that the current `EUKulele` version is 1.0.1 and `mamba` appears to be attempting to install 0.0.1, so I might try specifying the version explicitly. I usually also try removing strict channel priority via `conda config --describe channel_priority flexible`. Thank you @johanneswerner for helping with this! Hopefully we can reproduce and fix the problem.
