### 2020-12-03 AlexanderLabWHOI/EUKulele#24 @jcmcnch
@johanneswerner , I followed the install commands listed on readthedocs and referenced above.

```
 conda create -n EUKulele
 conda activate EUKulele
 conda install -c akrinos -c bioconda -c conda-forge EUKulele
```

@akrinos , tried making my conda flexible by using the following command that seems to be equivalent to what you're suggesting above (since the command you provided gave me an error): `conda config --set channel_priority true`

Unfortunately, this resulted in conda not being able to solve the environment - it ran for some ~4 hours without coming to a solution before I cancelled the job.

mamba also fails `mamba install -c akrinos -c bioconda -c conda-forge EUKulele==1.0.1` after changing channel priority with the following error:

```
Looking for: ['eukulele 1.0.1', 'python 3.6.0']

0 packages in https://conda.anaconda.org/akrinos/linux-64
3 packages in https://conda.anaconda.org/akrinos/noarch
33639 packages in https://conda.anaconda.org/bioconda/linux-64
21311 packages in https://conda.anaconda.org/bioconda/noarch
142297 packages in https://conda.anaconda.org/conda-forge/linux-64
43890 packages in https://conda.anaconda.org/conda-forge/noarch
18883 packages in https://repo.anaconda.com/pkgs/main/linux-64
2444 packages in https://repo.anaconda.com/pkgs/main/noarch
6637 packages in https://repo.anaconda.com/pkgs/r/linux-64
5025 packages in https://repo.anaconda.com/pkgs/r/noarch
47 packages in https://conda.anaconda.org/ursky/linux-64
0 packages in https://conda.anaconda.org/ursky/noarch
Encountered problems while solving.
Problem: package eukulele-1.0.1-pyh39e3cac_0 requires busco 4.0.6, but none of the providers can be installed
```

BUT, the good news is in the end updating conda did the trick, now it installs without issues. I suggest updating the install instructions to incorporate this and to use mamba as well, e.g.:

```
conda update -n base conda
conda create -n EUKulele
conda activate EUKulele
conda install -c conda-forge mamba
mamba install -c akrinos -c bioconda -c conda-forge EUKulele==1.0.1
```
