### 2020-12-03 AlexanderLabWHOI/EUKulele#24 @akrinos
Oops, looks like I provided the wrong command @jcmcnch, sorry about that!

Very interesting that updating `conda` changed that! The information that I found about the `mamba` error did all seem to point to something OS-related, but I really don't know why updating would have that effect. Maybe others have better insight!
