### 2020-12-07 AlexanderLabWHOI/EUKulele#24 @johanneswerner
Fascinating. Is the installation issue solved for now or should we look deeper into it?
