### 2020-12-08 #2817 @jcmcnch
@akrinos , I just ran the test suite and it seems to work find just by providing the yaml file as you describe, and everything seems to work - the errors mentioned above by Johannes seem to be fixed, except BUSCO generates no output. Is this expected? I checked the logs as recommended by the text printed to the screen but they were all empty. Here's the output I got:

```
(EUKulele) jesse@kraken:~/EUKulele-review/sample_EUKulele$ EUKulele --config curr_config.yaml
Running EUKulele with entries from the provided configuration file.
No BUSCO file specified/found; using argument-specified organisms and taxonomy for BUSCO analysis.
Setting things up...
Found database folder for reference_DIR in current directory; will not re-download.
Creating a diamond reference from database files...
Aligning to reference database...
Aligning sample sample_2...
Aligning sample sample_0...
Aligning sample sample_1...
Diamond process exited for sample sample_2.
Diamond process exited for sample sample_1.
Diamond process exited for sample sample_0.
Performing taxonomic estimation steps...
Performing taxonomic visualization steps...
Performing taxonomic assignment steps...
Performing BUSCO steps...
Configuring BUSCO...
Running busco with 1 simultaneous jobs...
[] is what is in BUSCO directory
BUSCO run either did not complete successfully, or returned no matches for sample sample_2 . Check busco_run log for details.
BUSCO run either did not complete successfully, or returned no matches for sample sample_0 . Check busco_run log for details.
BUSCO run either did not complete successfully, or returned no matches for sample sample_1 . Check busco_run log for details.
No BUSCO matches found for any sample.  Check BUSCO run log for details. Exiting...
EUKulele run complete
```
