### 2020-12-08 AlexanderLabWHOI/EUKulele#24 @jcmcnch
@johanneswerner it's solved at least from my end. Things are working well now and I'm proceeding with test-driving the software.
