### 2020-12-09 #2817 @akrinos
Hi @jcmcnch - sorry it has taken me a bit to get back to you. I have been trying to reproduce this, and haven't been able to. You are using the `sample_EUKulele` directory, right? This is what I expect to be printed:

```
Running busco with 1 simultaneous jobs...
['logs', 'short_summary.specific.eukaryota_odb10.sample_1.txt', 'run_eukaryota_odb10'] is what is in BUSCO directory
At least one BUSCO present in sample sample_1 but 250 missing.
At least one BUSCO present in sample sample_0 but 241 missing.
At least one BUSCO present in sample sample_2 but 245 missing.
```

Could you tell me (1) what `EUKulele --version` returns and (2) the contents of your sample directory (it should be `samples_MAGs` if you're using the tutorial) and `BUSCO` directory in the output folder via `ls`? Thanks!!
