### 2020-12-09 #2817 @jcmcnch
Hi @akrinos and other coauthors, again sorry for the delay in replying. I've had some time to properly "test drive" EUKulele, and now feel comfortable summarizing them as part of the review. I've noticed from your interactions with Johannes that this process seems quite interactive so I do hope we can discuss further in this thread. As I mentioned to Will at the beginning of this review process I'm somewhere closer to the naive end user and less of a software developer so will concentrate more on how I see your tool being used. These comments come from someone who is very interested in, but less knowledgeable about these "unsung" EUKs so it's kind of an outsider's view.

## Overall comments:

From the perspective of microbial oceanography, there seems to be a really strong cultural divide between people studying PROKs and people studying EUKs, despite the fact that the organisms in question interact in a larger system. So any effort to try and bridge this divide is really worthwhile scientifically and I think your tool and approach is a promising way to begin this effort. From my own perspective I had known about the MMETSP but was less confident to find/download the data. With EUKulele, having the database automatically downloaded is already very helpful and on top of that knowing that I'm getting a high-quality curated version of the MMETSP from experts in the field is really reassuring. I also really appreciate all the work that has gone into making EUKulele useful with multiple databases, bioinformatic methods, and providing visualizations.

My main feedback falls into two areas - 1) clarity of writing, code, and visualizations and 2) caveats applying your approach to mixed PROK/EUK metatranscriptomes. For 1), I will provide detailed comments further below, but a more general comment is that some broader concepts can be clarified for the benefit of those less familiar with eukaryotic work. For example, it took me a bit of time to understand what you mean by transcriptome. From the PROK side of the fence, transcriptomes are most often just something you map to your MAGs/contigs to get at expression but I recognize this is something quite different for EUKs - it's basically your metagenome. Clarifying this subtle distinction in the text might help people less familiar with your field understand this. I did see your warning in the readthedocs documentation about using EUK metagenomes which alludes to this issue but I think it could be further clarified and explained in a more prominent location. Otherwise, I think your already extensive documentation could be improved by some re-organization and re-focusing which I'll try to detail in the sections below. Also, I noted that metagenomes/MAGs were used somewhat interchangeably so this can also be clarified to explain what you mean.

Point 2) is a bit more getting at the real-world usage of the software, and a potential pitfall I see when your workflow is employed by a naive end user. It's related to Johannes' question about mixed PROK/EUK communities. To test this, I downloaded a transcriptome assembly from [this paper](https://www.nature.com/articles/s41597-019-0132-4), which can be found [here](https://drive.google.com/file/d/180RnRnmACbCwjY9Syfyua9iwHrXA72eu/view?usp=sharing) if you want to play with it yourself. Data were downloaded from IMG. From the phyloDB results generated by EUKulele, this is clearly a mixed PROK/EUK transcriptome:

![supergroup_transcripts](https://user-images.githubusercontent.com/5956604/101678259-87f90b80-3a12-11eb-873a-37247e7d8216.png)

My main concern is this is not reflected in the MMETSP results. Things that are clearly bacterial contigs (e.g. scaffold_10004_c1 which is a roseobacter) are annotated as EUKs (in this case, as a diatom). The tabular output (i.e. output/taxonomy_estimation/*taxonomy.out) would not give a user an idea that this is the case - the column for "max_pid" says 86.5% in this contig's case (100% for the phyloDB output file), so I wouldn't have assumed there was an issue unless I knew a priori that this sample would be a mixed EUK/PROK transcriptome assembly. How would you address this with your pipeline or in the paper/documentation? Do you mostly work with poly-A tailed transcriptomes in your own work where this would be less of an issue? Or is there a way you could think to address this? Could there be a pre-filtering step to split PROK and EUK? Or could another column be provided to the user to identify these potential issues?

## More specific comments

### Installation and testing:

This proceeded well after updating conda, but as I mentioned in the [bug report](https://github.com/AlexanderLabWHOI/EUKulele/issues/24) you may consider providing mamba as an alternative install method since it's much faster in general.

- [ ] Optional: provide mamba as alternative install method

The test suite worked without errors, except the BUSCO error I raised above. FYI I didn't install with pip as mentioned in the readthedocs section, but ran it as recommended:

`EUKulele --config curr_config.yaml`

In response to your question @akrinos :

Version:
```
(EUKulele) jesse@kraken:~/EUKulele-review/sample_EUKulele$ EUKulele --version
Running EUKulele with command line arguments, as no valid configuration file was provided.
The current EUKulele version is 1.0.1
```
Contents of directory:

```
(EUKulele) jesse@kraken:~/EUKulele-review/sample_EUKulele$ ls
busco_286409437.log   busco_327173586.log  EUKulele-env.yaml  path_test.txt    samples_MAGs
busco_2930947595.log  busco_downloads      free.csv           reference_DIR    tax-cutoffs.yaml
busco_3040169547.log  curr_config.yaml     output-test.txt    references_bins  test_out_23July
```

### Software usage:

Things were quite smooth. Database downloads proceeded as expected, and ran without errors on real-world MT data described above. I tested it with phyloDB and MMETSP using default settings on the MBARI bloom metatranscriptome assembly mentioned above.

Specific suggestions:

- [ ] I noticed dropbox is being used for the databases, which doesn't seem like a good long-term storage location. Suggest moving to a repo such as OSF/Zenodo where you can provide some additional documentation.
- [ ] Related to this, databases should have some README or info about the version, preprocessing, citation info, etc so this can be easily accessed and cited whenever someone uses your tool. I know this is mentioned in the readthedocs (i.e. Sarah's github repo) but it would be helpful to provide this as a plaintext README as well.
- [ ] The help file (output of `EUKulele --help`) is a little messy, I suggest reorganizing to move like parameters together (e.g. CPU and RAM usage), check for typos, and remove subroutines if it's not being used
- [ ] For visualizations, there are some things that could be improved, like making sure legends do not overlap with barplots, removing or improving uninformative labels ('OfInterest')
- [ ] Taxonomic output visualization files should have a numerical prefix that reflects the depth of each particular taxonomic level to make it easier to understand and scroll through
- [ ] Could busco log files be automatically moved to the output folder?

## Writing (paper and readthedocs):

- [ ] Please explain more explicitly why your pipeline is reproducible and flexible. You mention these two words and I can guess what you mean but I think it should stated more clearly.
- [ ] As mentioned above, the "warning note" about pitfalls using metagenomes should be more prominent in the documentation and explained further. I assume what you are referring to here is the difficulty of calling ORFs in genomic data when exons are interrupted by introns. Again, although this is obvious to you EUK folk it may help to explain in detail for the benefit of PROK people who tend to forget about these major wrinkles.
- [ ] I think the ordering of the readthedocs could be improved. Perhaps put "About" first, "Installing" second, and consider merging "Trying out EUKulele" with "EUKulele Quick Start" and putting both after "Installing"
- [ ] Consider putting the configuration file information in a section on its own and explain more what a yaml file is to the uninitiated and how it could be used to customize all parameters, not just taxonomic cutoffs.
- [ ] I agree with Johannes that more comparison with tools such as MEGAN and CCMetagen would be useful to orient the reader and place your work in a broader context
- [ ] There is some redundancy in the text on readthedocs, you could try to pare this down to make it easier to the reader, and mirror some of the bigger picture motivation and context here as well. For example, the first paragraph in "About" is more about functionality than the bigger picture and may be a bit redundant with the "Functionality" paragraph

Other than that, great job, thanks for sharing this powerful tool and your expert knowledge with the whole community. Looking forward to discussing more about the PROK/EUK mixed transcriptome assembly issue and how you see this being potentially addressed.

Jesse
