### 2020-12-11 #2817 @will-rowe
Thanks for another comprehensive review @jcmcnch - there is plenty for you to mull over @akrinos! I'm particularly interested in the second of their points re. (mis)reporting of prokaryotic annotations. @jcmcnch has given several helpful suggestions to address this; at a minimum I'd like a comment in the documentation. In my view, it would make sense for the user to first bin their MAGs (or alternative input) into EUK and PROK before running EUKulele - but this functionality would be great to have in your software.

Please let us know your responses to the second review. We are definitely well on track here. I also have a couple of things for you to address in the paper:
* Please give full author affiliations/addresses.
* The summary is quite long and has some large whitespace (not really your fault, just awkward figure placement by the renderer). In my opinion, the summary should be more of an abstract/single paragraph and followed up quickly by your statement of need. I feel this makes the paper easier to read and I can get to why your software is amazing more quickly. That being said, it is your paper and it fits our publication criteria.
* As a follow on point, it would be great to have a bit more detail on the specifics of your implementation. I would suggest taking what you already have (currently within your summary section) and fleshing it out under a new "implementation" section (or similar). Take a look at somer [recent JOSS papers](https://joss.theoj.org/) for inspiration.

Only the first of these points is a requirement from me!

Cheers,

Will
