### 2020-12-11 #2817 @akrinos
Thank you both for your very helpful comments! We are working on addressing the eukaryote/prokaryote mislabeling issue by generating a default database that contains the MMETSP taxonomy (the taxonomy we feel is better to use as a default in most cases for eukaryotic organisms) and also prokaryotic sequences and a domain level to distinguish between the two. We will include the MMETSP alone as an option, as we are a bit biased towards poly-A-selected samples for which you would likely only need to quickly check with a database like PhyloDB for whether contigs were preferentially mapping to bacteria, after which point the MMETSP would be the database of choice.

We will certainly adjust the organization of the summary section somewhat! For the affiliations, is adding the city/state/country names sufficient beyond what we have? It looks like that is what is included in the JOSS papers I looked at. More comments to come later today in regard to addressing more of the housekeeping issues with the repository as well. Thanks again!
