### 2020-12-13 #2817 @akrinos
@jcmcnch thanks so much again for your very thorough and informative review! I have responded to some of the points you raised below, and will complete the process of addressing them and provide a new release tomorrow.

In response to the major issue with the use of MMETSP resulting in prokaryotic sequences being labeled as eukaryotic, I have added a new database to the default options, a combination of [MarRef](https://mmp.sfb.uit.no/databases/marref/) and the MMETSP. This is to enable prokaryotes to be identified, but also to use our preferred eukaryotic sequences. Here is a comparison of the output for each database, MMETSP, PhyloDB, and MarRef, using the sample dataset that you provided:

![image](https://user-images.githubusercontent.com/22385511/102014280-2b079900-3d23-11eb-9464-68a47a1d1a61.png)

This also involves adding a separate "Domain" level to the MMETSP and MarRef. For now, I have made the software flexible, such that it will accept a number of labeling options from your database (Domain, Supergroup, Kingdom, Phylum, Class, Order, Family, Genus, Species, with the potential for more to be added), and arrange them based on expected taxonomic ordering. In the future, I plan to relabel the top level of PhyloDB to be "Domain" as well, since it makes more sense to use that label for the highest taxonomic level in PhyloDB, rather than using the MMETSP's "Supergroup" name.

So, the issue persists if you use the MMETSP on prokaryotic sequences - there tend to be spurious matches at the supergroup level, since the percent identity cutoff is quite low at this broad level, but we think it's important to retain the MMETSP option if you wish to only consider eukaryotic matches. However, we will add an additional warning to the documentation, and MarRef-MMETSP has become the default, to avoid this occurring naively. Note that when using the MMETSP reference, many more are "unclassified" at more specific taxonomic levels than when using PhyloDB.

**Installation and testing**

- We have included a note about using `mamba` to install `EUKulele` on the installation page
- As far as the BUSCO issue, one more question - you installed with `conda`?

**Software usage**

- We have added the newest database, which contains both eukaryotic and prokaryotic sequences, to Zenodo. The reason we didn't do this before is/was twofold (1) we need to discuss the matter with some of the owners of the databases (e.g. `PhyloDB`, which is currently hosted on Google Drive and has not been formally published) and (2) using Zenodo requires that an additional dependency, `zenodo-get`, be included, which somewhat complicates the `conda` install. However the newly-added database should work fine with Zenodo as of now.
-  In progress
- Subroutine is something we intend to be able to be used, so I've left that in - I will be tweaking the help more shortly and it will be updated in the new release
- I removed the legend title and made it so that the legend is outside of the plot
- This is a great suggestion! Numerical prefixes will also be included in the next release
- Do you mean to have the BUSCO log files specifically outside of the `log` folder?

Thanks again, and more soon!
