### 2020-12-13 11732b6
Fixing legend in visualizations; adding new database; adding clarifications to paper
