+++ paper/paper.bib
@@ -7,6 +7,30 @@
   url = {https://github.com/AlexanderLabWHOI/EUKulele}
 }
 
+@article{klemetsen2018mar,
+  title={{The MAR databases: development and implementation of databases specific for marine metagenomics}},
+  author={Klemetsen, Terje and Raknes, Inge A and Fu, Juan and Agafonov, Alexander and Balasundaram, Sudhagar V and Tartari, Giacomo and Robertsen, Espen and Willassen, Nils P},
+  journal={{Nucleic Acids Research}},
+  volume={46},
+  number={D1},
+  pages={D692--D699},
+  year={2018},
+  doi={https://doi.org/10.1093/nar/gkx1036},
+  publisher={Oxford University Press}
+}
+
+@article{marcelino2020ccmetagen,
+  title={{CCMetagen: comprehensive and accurate identification of eukaryotes and prokaryotes in metagenomic data}},
+  author={Marcelino, Vanessa R and Clausen, Philip TLC and Buchmann, Jan P and Wille, Michelle and Iredell, Jonathan R and Meyer, Wieland and Lund, Ole and Sorrell, Tania C and Holmes, Edward C},
+  journal={{Genome Biology}},
+  volume={21},
+  number={1},
+  pages={1--15},
+  year={2020},
+  publisher={BioMed Central},
+  doi={https://doi.org/10.1186/s13059-020-02014-2}
+}
+
 @article{Bolyen2018,
     author = {Bolyen, Evan and Rideout, Jai Ram and Dillon, Matthew R and Bokulich, Nicholas A and Abnet, Christian and Al-Ghalith, Gabriel A and Alexander, Harriet and Alm, Eric J and Arumugam, Manimozhiyan and Asnicar, Francesco and Bai, Yang and Bisanz, Jordan E and Bittinger, Kyle and Brejnrod, Asker and Brislawn, Colin J and Brown, C Titus and Callahan, Benjamin J and Caraballo-Rodriguez, Andres Mauricio and Chase, John and Cope, Emily and Silva, Ricardo Da and Dorrestein, Pieter C and Douglas, Gavin M and Durall, Daniel M and Duvallet, Claire and Edwardson, Christian F and Ernst, Madeleine and Estaki, Mehrbod and Fouquier, Jennifer and Gauglitz, Julia M and Gibson, Deanna L and Gonzalez, Antonio and Gorlick, Kestrel and Guo, Jiarong and Hillmann, Benjamin and Holmes, Susan and Holste, Hannes and Huttenhower, Curtis and Huttley, Gavin and Janssen, Stefan and Jarmusch, Alan K and Jiang, Lingjing and Kaehler, Benjamin and Kang, Kyo Bin and Keefe, Christopher R and Keim, Paul and Kelley, Scott T and Knights, Dan and Koester, Irina and Kosciolek, Tomasz and Kreps, Jorden and Langille, Morgan GI and Lee, Joslynn and Ley, Ruth and Liu, Yong-Xin and Loftfield, Erikka and Lozupone, Catherine and Maher, Massoud and Marotz, Clarisse and Martin, Bryan D and McDonald, Daniel and McIver, Lauren J and Melnik, Alexey V and Metcalf, Jessica L and Morgan, Sydney C and Morton, Jamie and Naimey, Ahmad Turan and Navas-Molina, Jose A and Nothias, Louis Felix and Orchanian, Stephanie B and Pearson, Talima and Peoples, Samuel L and Petras, Daniel and Preuss, Mary Lai and Pruesse, Elmar and Rasmussen, Lasse Buur and Rivers, Adam and {Michael S Robeson}, II and Rosenthal, Patrick and Segata, 
[…]
