+++ src/EUKulele/static/tax-cutoffs.yaml
@@ -3,3 +3,5 @@ genus: 80
 family: 65
 order: 50
 class: 30
+supergroup: 15
+domain: 5

[…]
