### 2020-12-13 bd7ae00
Fixing reference in paper and 1 more affiliation
