### 2020-12-13 e70f6c0
Fixing reference in paper and 1 more affiliation
