### 2020-12-13 3c47952
Fixing reference in paper and 1 more affiliation
