### 2020-12-15 be5ca34
Modifying documentation and adding versioning readout
