+++ docs/source/about.rst
@@ -2,7 +2,7 @@
 About ``EUKulele``
 ====================================
 
-``EUKulele`` is a python package that enables the rapid taxonomic annotation of metagenomic and metatranscriptomic data using a Last Common Ancestor (LCA) approach to align samples to a reference database. Assignments are chosen using either (a) the consensus annotation above a user-specified threshold, by default 75% identical predicted taxonomic annotations, or (b) the last conserved taxonomic level between the identified matches. ``EUKulele`` has been designed to be flexible and reproducible. In addition to estimating a taxonomic annotation for each putative transcript (“contig”), `EUKulele` can predict the taxonomic identity of eukaryotic Metagenome Assembled Genomes (MAGs) using the consensus of the contigs in each MAG. `EUKulele` also offers users a secondary taxonomic annotation approach limited only to sequenced reads annotated as core eukaryotic genes. 
+``EUKulele`` is a python package that enables the rapid taxonomic annotation of metagenomic and metatranscriptomic data using a Last Common Ancestor (LCA) approach to align samples to a reference database. Assignments are chosen using either (a) the consensus annotation above a user-specified threshold, by default 75% identical predicted taxonomic annotations, or (b) the last conserved taxonomic level between the identified matches. ``EUKulele`` has been designed to be flexible and reproducible. The software is flexible in that it allows any sequences the user is interested in to be added as a reference, and also provides the user the freedom to choose their own levels of taxonomic specificity and confidence in alignment matches. The pipeline is reproducible because the specific parameters used to run ``EUKulele`` and the output can be saved, containing information about which database was run using which version, allowing users to share metadata and easily defend their taxonomic annotation results. In addition to estimating a taxonomic annotation for each putative transcript (“contig”), `EUKulele` can predict the taxonomic identity of eukaryotic Metagenome Assembled Genomes (MAGs) using the consensus of the contigs in each MAG. `EUKulele` also offers users a secondary taxonomic annotation approach limited only to sequenced reads annotated as core eukaryotic genes. 
 
 A variety of curated protein databases are available to use with `EUKulele`, which are downloaded and formatted automatically. A custom local database can also be created by the user, using user-provided sequences and accompanying taxonomy. 
 
