### 2020-12-15 #2817 @akrinos
Again, thank you @jcmcnch and @will-rowe for your help!

As a follow-on:
**Writing (paper and readthedocs)**

- I have added a sentence of explanation in the about section of the readthedocs that further elaborates on reproducibility and flexibility
- I have added a note to the section of "Running `EUKulele`" that contains this warning explicitly saying that this is particularly important due to the presence of introns in eukaryotes. I have also added a link to this page of the documentation on the `EUKulele` landing page, such that users arriving at the documentation can be immediately aware of where the caveats are
- Though we recommend that users read the documentation, "EUKulele Quick Start" is meant as a quick landing page that could be used solely to invoke `EUKulele`, if a user was not interested in further invocation details. So we consider the Quick Start to be a separate resource from any tutorials or from the installation and about pages
- I have added a couple of lines to the database and configuration section about YAML files and how you would create one for the purposes of the taxonomic cutoffs. Since providing a full YAML config is not required to run `EUKulele`, I think that further specifics about YAML files beyond the beginner level are a little confusing to include in the main documentation. However, since users would need to create/modify a YAML file for the taxonomic cutoffs, I've clarified the documentation for that
- I have added reference to both of these tools in the paper and expanded that section a bit
- I have made some additional formatting changes to the readthedocs, but I expect that this will be a continually evolving process as we receive feedback from other users

As far as the issue of providing metadata for each of the databases, for now I am writing a file with each invocation of `EUKulele` that contains
- the name of the database
- the link from which the database was downloaded (if it was downloaded during that invocation)
- the date/time of invocation
This enables at least a record of what was used, and then to go to the documentation to learn more about the databases and cite those databases appropriately. In the future we will endeavor to provide something more direct.

I have just pushed a new release to both [`PyPI`](https://pypi.org/project/EUKulele/1.0.2/) and [`conda`](https://anaconda.org/akrinos/eukulele) containing all of the relevant code changes. Please let me know what you think of the various changes!
