### 2020-12-17 #2817 @jcmcnch
Hi @akrinos , thanks for your reply, this all looks great. I have just re-downloaded EUKulele using conda (and yes, with the BUSCO issue it was from a conda install but I will try again to make sure I get the same behaviour as before). I do have a few quick questions though:

1. Is there an easy way I confirm I've got the latest and greatest? After doing a clean install, `EUKulele --version` returns 1.0.1 which is the same as before so I'm not sure I'm getting the changes you've implemented
2. It looks like the readthedocs (latest) does not yet include the updates you mention above (e.g. mamba, warnings, usage instructions etc). Please clarify where I can find this info.
3. I can see the PDF above has been altered to include a lot more explanation and clarification. Is this the final version you want us to treat as a "response to reviewers"? BTW there are a few typos (e.g. prokaryotic spelled wrong, a couple references either missing or misspelled).
