### 2020-12-17 aa11d56
Fixing adding to wrong index.rst and lack of mamba instruction commit
