+++ paper/paper.bib
@@ -284,7 +284,7 @@
 @incollection{keegan2016mg,
   title={MG-RAST, a metagenomics service for analysis of microbial community structure and function},
   author={Keegan, Kevin P and Glass, Elizabeth M and Meyer, Folker},
-  booktitle={Microbial environmental genomics (MEG)},
+  booktitle={{Microbial Environmental Genomics (MEG)}},
   doi={10.1093/bioinformatics/btv351},
   pages={207--233},
   year={2016},
