+++ paper/paper.bib
@@ -319,9 +319,9 @@
 }
 
 @article{kent2002blat,
-  title={BLAT—the BLAST-like alignment tool},
+  title={{BLAT—the BLAST-like alignment tool}},
   author={Kent, W James},
-  journal={Genome research},
+  journal={{Genome Research}},
   doi={10.1101/gr.229202},
   volume={12},
   number={4},
@@ -331,9 +331,9 @@
 }
 
 @article{zhang2015tara,
-  title={The Tara Oceans project: new opportunities and greater challenges ahead},
+  title={{The Tara Oceans project: new opportunities and greater challenges ahead}},
   author={Zhang, Houjin and Ning, Kang},
-  journal={Genomics, proteomics \& bioinformatics},
+  journal={{Genomics, Proteomics \& Bioinformatics}},
   doi={10.1016/j.gpb.2015.08.003},
   volume={13},
   number={5},
@@ -345,7 +345,7 @@
 @article{tully2018reconstruction,
   title={The reconstruction of 2,631 draft metagenome-assembled genomes from the global oceans},
   author={Tully, Benjamin J and Graham, Elaina D and Heidelberg, John F},
-  journal={Scientific data},
+  journal={{Scientific Data}},
   doi={10.1038/sdata.2017.203},
   volume={5},
   pages={170203},
@@ -366,9 +366,9 @@
 }
 
 @article{buchfink2015fast,
-  title={Fast and sensitive protein alignment using DIAMOND},
+  title={{Fast and sensitive protein alignment using DIAMOND}},
   author={Buchfink, Benjamin and Xie, Chao and Huson, Daniel H},
-  journal={Nature methods},
+  journal={{Nature Methods}},
   doi={10.1038/nmeth.3176},
   volume={12},
   number={1},
@@ -378,18 +378,18 @@
 }
 
 @article{richter2020eukprot,
-  title={EukProt: a database of genome-scale predicted proteins across the diversity of eukaryotic life},
+  title={{EukProt: a database of genome-scale predicted proteins across the diversity of eukaryotic life}},
   author={Richter, Daniel J and Berney, Cedric and Strassert, Jurgen FH and Burki, Fabien and De Vargas, Colomban},
-  journal={BioRxiv},
+  journal={{BioRxiv}},
   doi={10.1101/2020.06.30.180687},
   year={2020},
   publisher={Cold Spring Harbor Laboratory}
 }
 
 @article{fell1992partial,
-  title={Partial rRNA sequences in marine yeasts: a model for identification of marine eukaryotes.},
+  title={{Partial rRNA sequences in marine yeasts: a model for identification of marine eukaryotes.}},
   author={Fell, JW and Statzell-Tallman, A and Lutz, MJ and Kurtzman, CP},
-  journal={Molecular marine biology and biotechnology},
+  journal={{Molecular Marine Biology and Biotechnology}},
   volume={1},
   number={3},
   pages={175--186},
@@ -397,9 +397,9 @@
 }
 
 @article{keeling2014marine,
-  title={The Marine Microbial Eukaryote Transcriptome Sequencing Project (MMETSP): illuminating the functional diversity of eukaryotic life in the oceans through transcriptome sequencing},
+  title={{The Marine Microbial Eukaryote Transcriptome Sequencing Project (MMETSP): illuminating the functional diversity of eukaryotic life in the oceans through transcriptome sequencing}},
[…]
