### 2020-12-17 #2817 @akrinos
Hi @jcmcnch, thanks so much for the feedback!

1. Unfortunately, I accidentally didn't update the version number on the `conda` side, but did on the `pip` side. I will fix that! But also, the easiest way to check is either to run `conda update`, or to check the Anaconda Cloud page. At the time that I'm writing this, our [page](https://anaconda.org/akrinos/eukulele) indicates that 1.0.1 was updates ~2 days ago, so that's probably the easiest way to check the version correspondence...but this is my fault for not relabeling as 1.0.2 on `conda`.
2. I didn't include the landing page warning accidentally (edited an old file by accident!) and have just committed the `mamba` section. The update on YAML and the introns were just 1-3 lines or so.
3. We would love to have that information included, however we know it is a little long for JOSS, so we are happy to also cut if needed. If it can be considered, that would be great, as we felt before that it would benefit from more description of the implementation. Sorry about that typo! I re-spell-checked the manuscript and that (prokaryotic) seemed to be the only spelling error...the missing citation was an unfortunate carry-over from renaming the citation. Thanks for checking both and please let me know if there are others I missed!

Thanks again!
