### 2021-01-05 #2817 @will-rowe
Hi @akrinos

Can you please check/fix those references. Looks like the MEGAN one might not need the `_4`

Once you have done this, please can you tag a new release and then archive it (with zenodo or similar). Then report back here with the DOI and version.
