### 2021-01-05 bea3737
Fixing references in paper
