+++ paper/paper.bib
@@ -15,7 +15,7 @@
   number={D1},
   pages={D692--D699},
   year={2018},
-  doi={https://doi.org/10.1093/nar/gkx1036},
+  doi={10.1093/nar/gkx1036},
   publisher={Oxford University Press}
 }
 
@@ -28,7 +28,7 @@
   pages={1--15},
   year={2020},
   publisher={BioMed Central},
-  doi={https://doi.org/10.1186/s13059-020-02014-2}
+  doi={10.1186/s13059-020-02014-2}
 }
 
 @article{Bolyen2018,
@@ -70,7 +70,7 @@
     number = {1},
     pages = {373},
     title = {{A global ocean atlas of eukaryotic genes}},
-    url = {https://doi.org/10.1038/s41467-017-02342-1},
+    url = {10.1038/s41467-017-02342-1},
     volume = {9},
     year = {2018}
 }
@@ -278,6 +278,7 @@
   booktitle={Functional Metagenomics: Tools and Applications},
   pages={65--74},
   year={2017},
+  doi={10.1007/978-3-319-61510-3_4},
   publisher={Springer}
 }
 
