### 2021-01-05 e7ac404
Adding requirement of bibtex 1.0.0 due to incompatibilities for docs building
