### 2021-01-05 #2817 @akrinos
Hi @will-rowe (and thank you @johanneswerner!) - I have fixed the DOI issues listed above, and published a release to Zenodo [here](https://zenodo.org/record/4419894#.X_SCFqpKg6U), with DOI 10.5281/zenodo.4419894 for version 1.0.2, which is also fully updated on Anaconda Cloud and PyPI. I left in the _4 for MEGAN, as that is the most specific DOI for the paper. Thank you so much for your help!
