### 2021-01-05 #2817 @will-rowe
Good work - thanks @akrinos. I'm afraid one more thing is needed from my end - can you make sure the zenodo release has an author list that matches the author list in your paper?
