### 2021-01-07 #2817 @akrinos
Hi @will-rowe, thanks and no problem! I couldn't figure out how to edit the author list before. I ended up having to modify it to be release 1.0.2b on Zenodo [here](https://zenodo.org/record/4422091#.X_Zfy6pKgUs); hopefully that's okay!
