### 2021-01-08 #2817 @arfon
@johanneswerner, @jcmcnch - many thanks for your reviews here and to @will-rowe for editing this submission. JOSS relies upon the volunteer efforts of folks like yourselves and we simply couldn't do this without you! ✨
